update-index: move add_cacheinfo() to read-cache.c
This moves the function add_cacheinfo() that already exists in
update-index.c to update-index.c, renames it add_to_index_cacheinfo(),
and adds an `istate' parameter.  The new cache entry is returned through
a pointer passed in the parameters.  This function can return three
values:

 - 0, when the file has been successfully added to the index;
 - ADD_TO_INDEX_CACHEINFO_INVALID_PATH, when the file does not exists;
 - ADD_TO_INDEX_CACHEINFO_UNABLE_TO_ADD, when the file could not be
   added to the index.

This will become useful in the next commit, when the three-way merge
will need to call this function.

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>