clone: teach --detach option
Teach "git clone" the "--detach" option, which leaves the cloned repo in
detached HEAD (like "git checkout --detach"). If the clone is not bare,
the remote's HEAD branch is also not created (bare clones always copy
all remote branches directly to local branches, so the branch is still
created in the bare case).

This is especially useful in the "submodule.propagateBranches" workflow,
where the submodule branch names match the superproject's branch names,
so it makes no sense to name the branches after the submodule's remote's
branches.

Signed-off-by: Glen Choo <chooglen@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>