clone, submodule update: check out branches
Teach "git submodule update" to update submodules by creating and
checking out the current superproject branch when
"submodule.propagateBranches=true". "git clone --recurse-submodules"
also learns this trick because it is implemented with "git submodule
update --recursive".

With "submodule.propagateBranches=true", submodules are cloned with
"--detach" so that they do not contain branches from their upstream.
This prevents conflicts between branch names from the superproject and
the branch names from the submodule's upstream. Arguably, "--detach"
should also be the default for "submodule.propagateBranches=false"
since it doesn't make sense to create a submodule branch when the
submodule is always expected to be in detached HEAD. But, to be
conservative, this commit does not change the behavior of
"submodule.propagateBranches=false".

"git submodule update" tries to create the branch as long as it is not
currently checked out, thus it will fail if the submodule has the
branch, but it is not checked out. This is fine because the main purpose
of "git submodule update" is to clone new submodules (which have no
branches, and will never have this problem). "git checkout" with
"submodule.propagateBranches" will cover the use case of recursively
checking out an existing branch.

Signed-off-by: Glen Choo <chooglen@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>